ai_services/requirements.txt
SpeechRecognition
PyAudio
python-socketio[client]
flask
flask-socketio
eventlet
pvporcupine
pvrecorder
openai-whisper
piper-tts
opencv-python
pyautogui
python-dotenv
mss
numpy
Pillow

